Allow terms and privacy to be configurable - Add option to set optional terms and optional privacy in config - Adjust the accept conditions checkbox when connecting wallet - Show Terms and Privacy links in menu if they exist
